Medium Changes
-Changes something in the config file that makes it so you can't
overscroll (so it doesn't bounce up and down when you scroll past in
iOS/show the blue stuff in android when you reach the top and bottom of
the page)
-Change the login page and added a proper external CSS style sheet.
Still currently working on a method to make the entire page fit no
matter what the phone screen size. I am currently enclosing the webpage
inside a div called wrapper and messing with the %width and minimum and
maximum pixel widths, along with the margins.