Pretty Stuff
I edited all of the icons to make them look less awful, changed the B’s
in the beginning (feel free to ignore the fact that the B isn’t
centered… I’ll work on it), and fixed the off-by-a-few-pixels thing
that was happening on the home screen. Like always, let me know what
looks stupid.