Started adding Talon UI
Modified Talon.html to grab the entire list of links from Parse. For
whatever reason, the query returns them completely unsorted, so I wrote
a quick, mildly efficient sorting algorithm which puts them in order
from newest to oldest. Then it displays the issue number and year for
the first five links in the array to the user with a view button.
There's also a more button to load the next five. However, neither of
these buttons have been configured yet.